Fixed a few lint/format/convention issues

+++ lib/curved_text.dart
@@ -26,23 +26,28 @@ class CurvedText extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    const maxRadius = 100000;
-    final desiredRadius = 1 / curvature.abs();
-    final radius = desiredRadius.abs() > maxRadius ? desiredRadius.sign * maxRadius : desiredRadius;
+    const maxRadius = 100000.0;
+    final desiredRadius = 1.0 / curvature.abs();
+    final radius = desiredRadius.abs() > maxRadius
+        ? desiredRadius.sign * maxRadius
+        : desiredRadius;
 
     final targetRadius = this.targetRadius;
-    final yDisp = targetRadius != null ? -curvature.sign * (targetRadius * targetRadius / radius) : 0;
+    final yDisp = targetRadius != null
+        ? -curvature.sign * (targetRadius * targetRadius / radius)
+        : 0.0;
 
     final yTranslate = curvature.isNegative ? -radius : radius;
-    final direction = curvature.isNegative ? Direction.counterClockwise : Direction.clockwise;
-    final startAngle = curvature.isNegative ? pi : 0;
+    final direction =
+        curvature.isNegative ? Direction.counterClockwise : Direction.clockwise;
+    final startAngle = curvature.isNegative ? pi : 0.0;
 
     return ArcText(
       direction: direction,
       radius: radius,
       text: text,
       textStyle: textStyle,
-      startAngle: startAngle.toDouble(),
+      startAngle: startAngle,
       startAngleAlignment: StartAngleAlignment.center,
       placement: Placement.middle,
       painterDelegate: (canvas, size, painter) {
